fix: 计费列改 numeric(20,4) 支持 sub-cent 文本/音频定价
解决「integer 计费列 vs sub-cent 定价」冲突（E2E 冒烟实测触发）：calculateCost
对文本（按 1M token）与音频（按秒）产出小数分（如 qwen-max 1000/500 token →
0.72 分），落库到 integer 列抛 22P02 invalid input syntax for type integer，
导致 /generate 文本与 /v1/chat/completions gateway 任何非整分成本都失败（视频/
图片为整分不受影响）。方案：credit ledger 与计费冗余列改 numeric(20,4)，而非
storage 边界取整（取整会让 ASR / fun-music 按秒计费损失真实收入）。

- schema 11 列 integer→numeric(20,4) mode:'number'（generation_records /
  canvas_assets / credit ledger / api_keys 共 4 文件）。mode:'number' 让 drizzle
  在 ORM 层把 numeric 映射为 number（.$type<number>() 只改 TS 类型不改运行时，
  实测分值列经 drizzle 读回仍是 string）。
- db.ts 注册 numeric(1700)→Number type parser（导出 numericTypeParser，测试
  helper 复用）：mode:'number' 只覆盖列读，但 admin 仓储的裸 SQL 聚合
  （coalesce(sum(...),0)，非列引用）绕过 ORM 映射，需驱动层 parser 统一覆盖。
- credit.repo.ts assertPositiveAmount 放宽为允许正小数（仍拒绝 0/负数/NaN）。
- admin 仓储去 ::int 截断保留 numeric 聚合精度（count(*)::int 计数保留）。
- 迁移 0037_billing_fractional_cents.sql（bare SQL，匹配 0034-0036 约定，
  int→numeric 无损上转，无 USING 子句）。
- E2E: fake 文本 usage 由 0/0 改 1000/500 token（注册赠送 1000 分覆盖 sub-cent
  debit），text 旅程新增断言 sub-cent（0.72 分）reserve→debit 落库——正是此前
  会 22P02 崩溃的值。

验收：typecheck / lint(改动文件) / build / boundaries 全绿；test server 547 /
worker 104 / billing 42 / client 376 / 全 packages 绿；test:db 130/0；e2e 8/8
幂等（×2）。仅 packages/events 1 处预存在失败（addConnection 返回类型漂移，
与本改动无关）。对应 TODO §四 计费条目已标记完成。

+++ packages/db/src/repositories/admin/gateway.ts
@@ -50,8 +50,8 @@ export async function listAdminGatewayClients(
         email: accounts.email,
         totalKeyCount: sql<number>`count(*)::int`,
         activeKeyCount: sql<number>`count(*) filter (where ${apiKeys.revokedAt} is null)::int`,
-        totalSpendCents: sql<number>`coalesce(sum(${apiKeys.totalSpendCents}), 0)::int`,
-        totalQuotaCents: sql<number | null>`case when count(*) filter (where ${apiKeys.quotaMaxCents} is null) > 0 then null else coalesce(sum(${apiKeys.quotaMaxCents}), 0)::int end`,
+        totalSpendCents: sql<number>`coalesce(sum(${apiKeys.totalSpendCents}), 0)`,
+        totalQuotaCents: sql<number | null>`case when count(*) filter (where ${apiKeys.quotaMaxCents} is null) > 0 then null else coalesce(sum(${apiKeys.quotaMaxCents}), 0) end`,
         lastKeyActivityAt: sql<Date | null>`max(${apiKeys.lastUsedAt})`,
       })
       .from(accounts)
@@ -127,7 +127,7 @@ export async function getAdminGatewayClientDetail(
         id: accounts.id,
         username: accounts.username,
         email: accounts.email,
-        creditBalanceCents: sql<number>`coalesce(${creditAccounts.availableCents}, 0)::int`,
+        creditBalanceCents: sql<number>`coalesce(${creditAccounts.availableCents}, 0)`,
       })
       .from(accounts)
       .leftJoin(creditAccounts, eq(creditAccounts.accountId, accounts.id))
@@ -137,8 +137,8 @@ export async function getAdminGatewayClientDetail(
       .select({
         totalKeyCount: sql<number>`count(*)::int`,
         activeKeyCount: sql<number>`count(*) filter (where ${apiKeys.revokedAt} is null)::int`,
-        totalSpendCents: sql<number>`coalesce(sum(${apiKeys.totalSpendCents}), 0)::int`,
-        totalQuotaCents: sql<number | null>`case when count(*) filter (where ${apiKeys.quotaMaxCents} is null) > 0 then null else coalesce(sum(${apiKeys.quotaMaxCents}), 0)::int end`,
+        totalSpendCents: sql<number>`coalesce(sum(${apiKeys.totalSpendCents}), 0)`,
+        totalQuotaCents: sql<number | null>`case when count(*) filter (where ${apiKeys.quotaMaxCents} is null) > 0 then null else coalesce(sum(${apiKeys.quotaMaxCents}), 0) end`,
         lastKeyActivityAt: sql<Date | null>`max(${apiKeys.lastUsedAt})`,
       })
       .from(apiKeys)
@@ -146,7 +146,7 @@ export async function getAdminGatewayClientDetail(
     getDb()
       .select({
         gatewayCalls: sql<number>`count(*)::int`,
-        gatewaySpendCents: sql<number>`coalesce(sum(${generationRecords.totalPriceCents}), 0)::int`,
+        gatewaySpendCents: sql<number>`coalesce(sum(${generationRecords.totalPriceCents}), 0)`,
       })
       .from(generationRecords)
       .where(and(eq(generationRecords.accountId, accountId), sql`input_params->>'source' = 'gateway'`)),
